requirements.txt
Django
psycopg2
djangorestframework
requests
pandas
openpyxl